adding new module action class

+++ module/Engine/Engine.cs
@@ -2,5 +2,7 @@ using UnityEngine;
 using System.Collections;
 
 public class Engine : Module {
-  public int thrust;
+  public float thrust;
+
+  
 }
